changes and bugfixes all around

+++ src/pages/errors/NotFound.tsx
@@ -1,30 +1,46 @@
-import { Container, Row, Col, Alert, Button } from 'react-bootstrap';
+import { Button } from 'react-bootstrap';
 import '../../css/ErrorPages.css';
+import { fetchWithAuth } from '../../components/FetchWithAuth';
+import { useNavigate } from 'react-router-dom';
 
 // Define a página de erro 404
 export const NotFound = () => {
+  const navigate = useNavigate();
 
   // Remove o token do localStorage e redireciona para a página de login
-  const returnToLogin = () => {
-    localStorage.clear();
-    window.location.href = '/';
-  }
+  const returnToLogin = async () => {
+		try {
+			const token = localStorage.getItem('token');
+			const response = await fetchWithAuth('Authentication/Logout', {
+				method: 'POST',
+				headers: {
+					'Content-Type': 'application/json',
+				},
+				body: JSON.stringify({ token }),
+			});
+			if (response.ok) {
+				localStorage.clear();
+				navigate('/');
+			} else {
+				console.error('Erro:', response);
+			}
+		} catch (error) {
+			console.error('Erro:', error);
+		}
+
+	};
 
   return (
     <div className='background'>
-      <Container className="d-flex align-items-center justify-content-center" style={{ height: '100vh' }}>
-        <Row className="text-center">
-          <Col md={7} sm={8} xs={10}>
-            <Alert variant="light" className='alert'>
-              <img className='logo-img' src="/logo_login.png" alt="logo" />
-              <p>
-                Funcionalidade não disponível. Contacte o administrador do sistema.
-              </p>
-              <Button onClick={returnToLogin} className='link-button'>Voltar para o login</Button>
-            </Alert>
-          </Col>
-        </Row>
-      </Container>
+      <div className="error-container">
+        <form className="form-error">
+          <img className='logo-img' src="/logo_login.png" alt="logo" />
+          <p style={{ color: 'white' }}>
+            Funcionalidade não disponível. Contacte o administrador do sistema.
+          </p>
+          <Button style={{ marginTop: 20 }} onClick={returnToLogin} variant='outline-light'>Voltar para o login</Button>
+        </form>
+      </div>
     </div>
   );
 };
